// frontend/src/api/model/neighborhoodDTO.ts
/**
 * PropTech API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GeoJsonGeometry } from './geoJsonGeometry';


export interface NeighborhoodDTO { 
    id?: string;
    name?: string;
    polygon?: GeoJsonGeometry;
}