// frontend/src/api/model/propertySearchRequest.ts
/**
 * PropTech API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GeoJsonGeometry } from './geoJsonGeometry';


/**
 * Parámetros de búsqueda avanzada de inmuebles
 */
export interface PropertySearchRequest { 
    polygon?: GeoJsonGeometry;
    lat?: number;
    lng?: number;
    /**
     * Radio en metros (usado con lat/lng cuando no hay polygon)
     */
    radius?: number;
    minPrice?: number;
    maxPrice?: number;
    minRooms?: number;
    page?: number;
    size?: number;
}