// frontend/src/api/model/geoJsonGeometry.ts
/**
 * PropTech API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Objeto GeoJSON de geometría (Polygon o MultiPolygon)
 */
export interface GeoJsonGeometry { 
    type?: GeoJsonGeometry.TypeEnum;
    coordinates?: Array<any>;
}
export namespace GeoJsonGeometry {
    export const TypeEnum = {
        Polygon: 'Polygon',
        MultiPolygon: 'MultiPolygon'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}